ci: create a clone of 'latest' image
*NOTE: This commit shall not be merged in master!*

This is just a clone of the 'latest' image.

For the context:

The ceph-ansible upstream CI runs severals tests, including a
'idempotency/handlers' test. It means the playbook is run a first time
and then a second time with an other container image to ensure the
handlers run properly and the containers are well restarted.
But it can causes issues, for instance, I've hit the case where `latest`
image ships ceph 12.2.3 while the `stable-3.0` (which is the image used for
the second). ships ceph 12.2.2.

The goal of this test is not to verify we can upgrade from a specific
version to another specific version but to ensure handlers are working.

We just need to have a container image which has a different id for the
upstream CI, the goal is to have the same content in container image but
a different image id in the registry since the test relies on image id
to decide whether the container should be restarted.

*NOTE: This commit shall not be merged in master!*

Signed-off-by: Guillaume Abrioux <gabrioux@redhat.com>