replace: Adding GENERIC_ARCH
When building an x86_64 based distro, ARCH=x86_64 but some calls requests ARCH="amd64" as per the github release URLs.
This lead to a dilemna where in the same distro, some code requests ARCH=x86_64 and some ARCH=amd64.

This is where GENERIC_ARCH is made for.
- If ARCH=x86_64 then GENERIC_ARCH is set to amd64
- If ARCH=amd64  then GENERIC_ARCH is set to amd64

amd64 is then the generic arch name of both x86_64 & amd64 and could be use everywhere when you need to reference it.

This first patch is about adding the definition of GENERIC_ARCH and implement it for CONFD, ETC, FOREGO & KUBECTL.