Build Ubuntu images for supported ceph versions
Since Ubuntu is the main image supported by ceph-container, for the
switch to the new project structure, start with enabling all Ubuntu
builds that previously existed.

Build luminous, kraken, and jewel for 16.04
Also build jewel for 14.04

kraken and jewel codebases have many files that differed from the
luminous source used as the source base for the new structure. Any
differing files are added as overrides in jewel/kraken. For future
versions of ceph, we should try to use the source base for all releases.

To speed up builds, a make parallel-build target was added to the
makefile. It uses make's `--jobs` flag with no arguments so that make
will intelligently choose the number of threads to use.

This commit also removes all the links and unnecessary files.

Original commit by: Blaine Gardner <blaine.gardner@suse.com>
Modified by Erwan Velu to use git mv instead of cp/rm to keep the tracability